refactor(linter/plugins): move token-related `SourceCode` methods into separate file (#14770)
Pure refactor. Move methods of `SourceCode` related to tokens into a separate file, to make main `source_code.js` file less crowded.
